support non root user on vscode

+++ src/common/scripts/docker-entrypoint.sh
@@ -29,47 +29,51 @@ if [ ! -f /etc/init-done ]; then
         echo "----------------------------------------"
     fi
 
-    if [ "$DISABLE_DESKTOP" = "true" ]; then
-        disable_service /etc/supervisord.d/novnc.ini
-        disable_service /etc/supervisord.d/xrdp.ini
-    fi
-    if [ "$DISABLE_TERMINAL" = "true" ]; then
-        disable_service /etc/supervisord.d/butterfly.ini
-    fi
-    if [ "$DISABLE_FILER" = "true" ]; then
-        disable_service /etc/supervisord.d/filebrowser.ini
-    fi
-    if [ "$DISABLE_SSH" = "true" ]; then
-        disable_service /etc/supervisord.d/sshd.ini
-    fi
-    if [ "$DISABLE_RDP" = "true" ]; then
-        disable_service /etc/supervisord.d/xrdp.ini
-    fi
-    if [ "$DISABLE_CODE" = "true" ]; then
-        disable_service /etc/supervisord.d/code.ini
-    fi
+    if [ -d /etc/xrdp ]; then
+        if [ "$DISABLE_DESKTOP" = "true" ]; then
+            disable_service /etc/supervisord.d/novnc.ini
+            disable_service /etc/supervisord.d/xrdp.ini
+        fi
+        if [ "$DISABLE_TERMINAL" = "true" ]; then
+            disable_service /etc/supervisord.d/butterfly.ini
+        fi
+        if [ "$DISABLE_FILER" = "true" ]; then
+            disable_service /etc/supervisord.d/filebrowser.ini
+        fi
+        if [ "$DISABLE_SSH" = "true" ]; then
+            disable_service /etc/supervisord.d/sshd.ini
+        fi
+        if [ "$DISABLE_RDP" = "true" ]; then
+            disable_service /etc/supervisord.d/xrdp.ini
+        fi
+        if [ "$DISABLE_CODE" = "true" ]; then
+            disable_service /etc/supervisord.d/code.ini
+        fi
 
-    # change sshd port
-    if [ "$SSH_PORT" != "" ]; then
-        sed -i "s/^#Port 22/Port $SSH_PORT/g" /etc/ssh/sshd_config
-    fi
+        # change sshd port
+        if [ "$SSH_PORT" != "" ]; then
+            sed -i "s/^#Port 22/Port $SSH_PORT/g" /etc/ssh/sshd_config
+        fi
 
-    # disable ssh password login
-    if [ "$DISABLE_SSH_PASSWORD_LOGIN" = "true" ]; then
-        sed -i "s/^PasswordAuthentication yes/PasswordAuthentication no/g" /etc/ssh/sshd_config
-    fi
+        # disable ssh password login
+        if [ "$DISABLE_SSH_PASSWORD_LOGIN" = "true" ]; then
+            sed -i "s/^PasswordAuthentication yes/PasswordAuthentication no/g" /etc/ssh/sshd_config
+        fi
 
-    # insert ssh public key
-    if [ "$SSH_KEY" != "" ]; then
-        mkdir $HOME/.ssh
-        chmod 700 $HOME/.ssh
-        echo $SSH_KEY >> $HOME/.ssh/authorized_keys
-        chmod 600 $HOME/.ssh/authorized_keys
-    fi
+        # insert ssh public key
+        if [ "$SSH_KEY" != "" ]; then
+            mkdir $HOME/.ssh
+            chmod 700 $HOME/.ssh
+            echo $SSH_KEY >> $HOME/.ssh/authorized_keys
+            chmod 600 $HOME/.ssh/authorized_keys
+        fi
 
-    # change rdp port
-    if [ "$RDP_PORT" != "" ]; then
-        sed -i "s/^port=3389/port=$RDP_PORT/g" /etc/xrdp/xrdp.ini
+        # change rdp port
+        if [ "$RDP_PORT" != "" ]; then
+            sed -i "s/^port=3389/port=$RDP_PORT/g" /etc/xrdp/xrdp.ini
+        fi
+
+        echo "VNC_RESOLUTION=$VNC_RESOLUTION" >> $HOME/.bashrc
     fi
 
     # initializing nginx
@@ -97,17 +101,10 @@ EOF
     sed -i "s/8080/$PORT/g" /etc/nginx/nginx.conf
     # initializing nginx done
 
-    echo "VNC_RESOLUTION=$VNC_RESOLUTION" >> $HOME/.bashrc
     if [ "DOCKER_HOST" != "" ]; then
         echo "export DOCKER_HOST=$DOCKER_HOST" >> $HOME/.bashrc
     fi
 
-    # post hook (once)
-    if [ "$POST_HOOK_ONCE" != "" ]; then
-        echo "---- post hook once : $POST_HOOK_ONCE ----"
-        source $POST_HOOK_ONCE || exit 1
-        echo "----------------------------------------"
-    fi
     mkdir $STARTUPDIR/tmp
     chmod 777 $STARTUPDIR/tmp
 
@@ -149,22 +146,28 @@ EOF
         fi
 
         cp -a /root/ /home/$USER
-        rm -rf /root/.ssh
         useradd $USER -u $USER_ID -d /home/$USER
         chown -R $USER:$USER /home/$USER
         echo "${USER}:${PASSWORD}" | chpasswd
-        chown -R $USER:$USER /var/lib/filebrowser
-        sed -i "s#-r /root#-r /home/$USER#g" /etc/supervisord.d/filebrowser.ini
-        sed -i "s/^username=ask$/username=$USER/g" /etc/xrdp/xrdp.ini
-        sed -i "s#sql:/root#sql:/home/$USER#g" /home/$USER/.mozilla/firefox/ygz7f6y1.myprofile/pkcs11.txt
-        sed -i "s#/root#/home/$USER#g" /home/$USER/.config/xfce4/xfconf/xfce-perchannel-xml/xfce4-desktop.xml
-        sed -i "s#/root#/home/$USER#g" /home/$USER/.config/xfce4/xfconf/xfce-perchannel-xml/xfce4-panel.xml
 
-        for i in ttyd.ini code.ini filebrowser.ini novnc.ini; do
-            echo "user=$USER" >> /etc/supervisord.d/$i
-            echo "directory=/home/$USER" >> /etc/supervisord.d/$i
-            echo "environment=HOME=\"/home/$USER\"" >> /etc/supervisord.d/$i
-        done
+        if [ -d /etc/xrdp ]; then
+            rm -rf /root/.ssh
+            chown -R $USER:$USER /var/lib/filebrowser
+            sed -i "s#-r /root#-r /home/$USER#g" /etc/supervisord.d/filebrowser.ini
+            sed -i "s/^username=ask$/username=$USER/g" /etc/xrdp/xrdp.ini
+            sed -i "s#sql:/root#sql:/home/$USER#g" /home/$USER/.mozilla/firefox/ygz7f6y1.myprofile/pkcs11.txt
+            sed -i "s#/root#/home/$USER#g" /home/$USER/.config/xfce4/xfconf/xfce-perchannel-xml/xfce4-desktop.xml
+            sed -i "s#/root#/home/$USER#g" /home/$USER/.config/xfce4/xfconf/xfce-perchannel-xml/xfce4-panel.xml
+
+            for i in ttyd.ini filebrowser.ini novnc.ini; do
+                echo "user=$USER" >> /etc/supervisord.d/$i
+                echo "directory=/home/$USER" >> /etc/supervisord.d/$i
+                echo "environment=HOME=\"/home/$USER\"" >> /etc/supervisord.d/$i
+            done
+        fi
+        echo "user=$USER" >> /etc/supervisord.d/code.ini
+        echo "directory=/home/$USER" >> /etc/supervisord.d/code.ini
+        echo "environment=HOME=\"/home/$USER\"" >> /etc/supervisord.d/code.ini
 
         if [ "$ENABLE_SUDO" = "true" ]; then
             echo "$USER	ALL=(ALL)	NOPASSWD: ALL" >> /etc/sudoers
@@ -174,15 +177,23 @@ EOF
         sed -i "s/^username=ask$/username=root/g" /etc/xrdp/xrdp.ini
     fi
 
+    # post hook (once)
+    if [ "$POST_HOOK_ONCE" != "" ]; then
+        echo "---- post hook once : $POST_HOOK_ONCE ----"
+        source $POST_HOOK_ONCE || exit 1
+        echo "----------------------------------------"
+    fi
     touch /etc/init-done
 else
     echo "skip initializing"
 fi
 
-if [ "$TTYD_OPTS" = "" ]; then
-    TTYD_OPTS='-P 30'
+if [ -d /etc/xrdp ]; then
+    if [ "$TTYD_OPTS" = "" ]; then
+        TTYD_OPTS='-P 30'
+    fi
+    export TTYD_OPTS
 fi
-export TTYD_OPTS
 
 # post hook
 if [ "$POST_HOOK" != "" ]; then
